+++ hamza-client/src/modules/products/components/product-review/index.tsx
@@ -16,7 +16,7 @@ import {
     StackDivider,
     Box,
 } from '@chakra-ui/react';
-import { averageRatings, reviewResponse, reviewCounter } from '@lib/data';
+import { averageRatings, getAllProductReviews, reviewCounter } from '@lib/data';
 
 const MEDUSA_SERVER_URL =
     process.env.NEXT_PUBLIC_MEDUSA_BACKEND_URL || 'http://localhost:9000';
@@ -54,7 +54,7 @@ const ProductReview: React.FC<ProductReviewProps> = ({
             // API calls remain the same
             const averageRatingResponse = await averageRatings(product.id);
             const reviewCountResponse = await reviewCounter(product.id);
-            const reviewsResponse = await reviewResponse(product.id);
+            const reviewsResponse = await getAllProductReviews(product.id);
 
             setAverageRating(averageRatingResponse);
             setReviewCount(reviewCountResponse);
